installer/requirements.txt
boto==2.49.0
boto3==1.17.49
botocore==1.20.49
aws-cdk.assets==1.101.0
aws-cdk.aws-backup==1.101.0
aws-cdk.aws-certificatemanager==1.101.0
aws-cdk.aws-cloudformation==1.101.0
aws-cdk.aws-cloudwatch==1.101.0
aws-cdk.aws-cloudwatch-actions==1.101.0
aws-cdk.aws-config==1.101.0
aws-cdk.aws-ec2==1.101.0
aws-cdk.aws-efs==1.101.0
aws-cdk.aws-directoryservice==1.101.0
aws-cdk.aws-elasticloadbalancing==1.101.0
aws-cdk.aws-elasticloadbalancingv2==1.101.0
aws-cdk.aws-elasticsearch==1.101.0
aws-cdk.aws-events==1.101.0
aws-cdk.aws-eventschemas==1.101.0
aws-cdk.aws-iam==1.101.0
aws-cdk.aws-kms==1.101.0
aws-cdk.aws-lambda==1.101.0
aws-cdk.aws-logs==1.101.0
aws-cdk.aws-s3==1.101.0
aws-cdk.aws-s3-assets==1.101.0
aws-cdk.aws-secretsmanager==1.101.0
aws-cdk.aws-sns==1.101.0
aws-cdk.aws-sns-subscriptions==1.101.0
aws-cdk.aws-ssm==1.101.0
aws-cdk.cloudformation-include==1.101.0
aws-cdk.core==1.101.0
aws-cdk.custom-resources==1.101.0
aws-cdk.region-info==1.101.0
aws-cdk.yaml-cfn==1.101.0
colored==1.4.2
requests==2.25.1
requests-aws4auth==1.0.1
PyYAML==5.4.1
